requirements.txt
torch>=2.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
pyyaml>=6.0
